refactor: rename date-utils to utils/date
