+++ bin/broadcast-reminder.ts
@@ -3,7 +3,7 @@ import { Result } from 'oxide.ts';
 import { getDiscordClient } from '../src/clients';
 import { formatReminderMessage, getReminderByTime, removeReminders } from '../src/commands/reminder/utils';
 import { getReminderChannel } from '../src/commands/serverSettings/utils';
-import { getCurrentUnixTime } from '../src/utils/date-utils';
+import { getCurrentUnixTime } from '../src/utils/date';
 import { loadEnv } from '../src/utils/load-env';
 import { logger } from '../src/utils/logger';
