+++ scripts/delete-guild-commands.ts
@@ -1,6 +1,6 @@
 import { Result } from 'oxide.ts';
 import { deployGuildCommands } from '../src/commands/deploy-command';
-import { getCurrentUnixTime } from '../src/utils/date-utils';
+import { getCurrentUnixTime } from '../src/utils/date';
 import { loadEnv } from '../src/utils/load-env';
 import { logger } from '../src/utils/logger';
