+++ scripts/deploy-guild-commands.ts
@@ -1,7 +1,7 @@
 import { Result } from 'oxide.ts';
 import { commandList, contextMenuCommandList } from '../src/commands';
 import { deployGuildCommands } from '../src/commands/deploy-command';
-import { getCurrentUnixTime } from '../src/utils/date-utils';
+import { getCurrentUnixTime } from '../src/utils/date';
 import { loadEnv } from '../src/utils/load-env';
 import { logger } from '../src/utils/logger';
